haste-worker-farm: pool.kill() method will call worker.kill() only if it's not dead already

+++ packages/haste-worker-farm/src/pool.js
@@ -48,7 +48,9 @@ module.exports = class Pool {
   }
 
   kill() {
-    this.ending = true;
-    this.workers.forEach(worker => worker.kill());
+    if (!this.ending) {
+      this.ending = true;
+      this.workers.forEach(worker => worker.kill());
+    }
   }
 };
